requirements.txt
PyQt5>=5.15
dnspython>=2.3
